HW-5: Basic Functional Implemented (no fetching data)

+++ evevseevPW-HomeProject/evevseevPW-HomeProject/Views/Extensions/UIView+Pin.swift
@@ -161,14 +161,14 @@ extension UIView {
     }
 
     @discardableResult
-    func setWidth(_ const: Int) -> NSLayoutConstraint {
+    func setWidth(_ const: CGFloat) -> NSLayoutConstraint {
         let constraint = widthAnchor.constraint(equalToConstant: CGFloat(const))
         constraint.isActive = true;
         return constraint;
     }
 
     @discardableResult
-    func setHeight(_ const: Int) -> NSLayoutConstraint {
+    func setHeight(_ const: CGFloat) -> NSLayoutConstraint {
         let constraint = heightAnchor.constraint(equalToConstant: CGFloat(const))
         constraint.isActive = true;
         return constraint;
